pyproject.toml
[project]
name = "o-gym"
version = "0.1.0"
description = "LeetCode, HackerRank, etc."
authors = [
    { name = "Roman Maliushkin" }
]
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0"
]

[build-system]
requires = [
    "setuptools>=61.0"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []